Backend/scheduler/requirements.txt
celery>=5.3.6
redis>=5.0.1
python-dotenv>=1.0.0
requests>=2.31.0
supabase>=2.3.0
openai>=1.12.0
psycopg2-binary>=2.9.9
